// generated/src/app/api-models/case-action.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CaseAction {
  seqCaseId: number;
  seqCaseReasonId: number;
  seqActionId: number;
  sequenceNumber: number;
  requestor: string;
  requestDate: Date;
  assignedRep: string;
  xrefTypeCsCode: string;
  xrefNumber: string;
  dueDate: Date;
  statusCsCode: string;
  statusDate: Date;
  request: string;
  respondToEntity: string;
  seqCallerId: number;
  responseMethodCsCode: string;
  responseLetterId: string;
  responseDate: Date;
  response: string;
  userDefined1: string;
  userDefined2: string;
  userDefinedDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}